Remove support for installing qemu-static from latest Ubuntu repo
I added this in times when Travis provided only ancient Ubuntu version
with ancient qemu package on which apk-tools didn't work. Since there's
an Ubuntu release hard-coded in the script, it caused problems when
Ubuntu removed repositories for the used release.

The times changed and this is not needed anymore, so let's get rid of
it.